Refactor entity reading by introducing EntityDataObject
This implementation of DataObject contains a reference to
the root entity built by EntityBuilderManager.

That should make it possible to data-object to contain values
which are not attributes of any entity.

Motivation behind this is to instantiate entity queries
which may be joined to table expressions containing GROUP BY -clause.